+++ setup/MacBootstrapSetup/Package.swift
@@ -3,16 +3,14 @@
 import PackageDescription
 
 let package = Package(
-    name: "MacBootstrapAgent",
+    name: "MacBootstrapSetup",
     platforms: [
         .macOS(.v13)
     ],
     products: [
-        .executable(name: "MacBootstrapAgent", targets: ["MacBootstrapAgent"]),
         .executable(name: "MacBootstrapSetup", targets: ["MacBootstrapSetup"])
     ],
     targets: [
-        .executableTarget(name: "MacBootstrapAgent"),
         .executableTarget(name: "MacBootstrapSetup")
     ]
 )